% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.enum-package.R
\docType{package}
\name{murlib.enum-package}
\alias{murlib.enum}
\alias{murlib.enum-package}
\title{murlib.enum: Super-Simple Enums}
\description{
Enums are strings.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
